% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barley.R
\docType{data}
\name{barley}
\alias{barley}
\title{barley}
\format{one dataset with 7 variables
\describe{
 \item{Rep}{Replicate,levels with 6}
 \item{RowBlk}{levels with 30}
 \item{ColBlk}{levels with 30}
 \item{Row}{levels with 10}
 \item{Column}{levels with 15}
 \item{Variety}{levels with 25}
 \item{yield}{a numeric vector}
}}
\usage{
data(barley)
}
\description{
barley dataset from asreml package.
}
\keyword{datasets}
